add hb_log_tcp

+++ src/hb_cache/hb_cache.c
@@ -22,7 +22,7 @@ hb_result_t hb_cache_available()
     return HB_SUCCESSFUL;
 }
 //////////////////////////////////////////////////////////////////////////
-hb_result_t hb_cache_initialize( const char * _uri, uint32_t _port )
+hb_result_t hb_cache_initialize( const char * _uri, uint16_t _port )
 {
     redisContext * c = redisConnect( _uri, _port ); //6379
